sxmo_bluetoothmenu.sh: allow to select the default controller
Being able to select between multiple controller become necessary
when you connect a bluetooth dongle to any devices wich already have an
internal bluetooth hardware. Atm we use the default controller, which is
completely random, and based on which one the daemons detected first.

To do so, we re-implement most of the bluetoothctl command using expect.

It is lighter and smarter than spaming the bluetoothctl commands. It
also simplify the scanning jobs, because we can do both scanning and
connecting with the same bluetoothctl interactive client.
And it also become necessary to allow the users to select the controller they
want. Because the bluetoothctl command only allow to run one command. We can not
chain "select $ctn" and "connect $dev" by example.

Also contains some minor code improvements.