Fix bluetooth script not being able to do things?
This is an old commit I got, and that was planned to land with the
switch to expect in this bluetooth script. But for some reason, I forgot
to include it...

sxmo_terminal.sh: make stdin/stdout available to sub-commands

We duplicate stdin to 4, and stdout to 5, so that the subprocess can use
`<&4 >&5` to use stdin/stdout of the sxmo_terminal.sh process.