sxmo_timer.sh: improve stopwatch mode
The stopwatch mode is a simple stop watch. In display a timer that
increase over time. It is very usefull to time recipes by example. No
keybinds exists to control it. An emergent usage is to press <ret>
to save timer moments, and keep displaying the current time on the last
line.

In case you want to reset the stopwatch, you have to stop the script
and restart another one.

This patch make the <ret> keybind to just reset the stopwatch. We keep
displaying the previous times on previous lines.

This patch also try hard to keep a correct history, even when the user
press other keys than <ret>. To do so, we move the cursor up after
<ret>, we re-display this line completely, and finaly we move down.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Aren Moynihan <aren@peacevolution.org>