tailtextlog: only attempt to focus a window when we're going to open a terminal window
This fixes a bug where a terminal would be launched, which would then
detect and focus its self, instead of tailing the log file. This would
result in the terminal window closing as soon as soon as it opened, and
focused it's self, which made it impossible to read logs.

Fixes: 8894f2a4 Try to simplify the terminal wrapping
Signed-off-by: Aren Moynihan <aren@peacevolution.org>
Signed-off-by: Willow Barraco <contact@willowbarraco.fr>