sxmo_jobs.sh: add ability to manage jobs by process group
If we want to run a shell pipeline as a job, we need to be able to kill
all processes. In order to do this, when we are starting a "group job",
we create a new session with setsid, and store the session id in
"$ROOT/id.group". Then when stopping a process, we can kill the entire
group if that file exists.

Signed-off-by: Aren Moynihan <aren@peacevolution.org>