sxmo_status_led: account for the brightness of multicolor leds
If the multi_intensity has non-zero values for leds, but the brightness
of the multicolor led is zero before a blink, we would leave the
brightness set to full, and the multi_intensity file unchanged.

On the PinePhone with the leds_group_multicolor driver enabled, this
resulted in the led being white after boot + the first blink. It starts
with brightness=0 and multi_intensity="1 1 1".

Signed-off-by: Aren Moynihan <aren@peacevolution.org>