sxmo_wm: reverse dpms logic
We used parameter 'sxmo_wm.sh dpms on' to turn the screen OFF and 'off'
to turn the screen ON, this is very counter intuitive and not in line with
what xset, wlr-randr and others do. This patch corrects it.

It also removes the sway-specific handling in favour of wlroots-generic
handling, but that does introduce wlr-randr as a new dependency.