sxmo_wm.sh: switch back to powering outputs instead of enabling them
Previously we relied on swaymsg to toggle the powering mode on Sway.
Recently we switched to wlr-randr to enable/disable the heads.

The problems is that it causes issues with some applications, even
crashes of Sway in case mpv is started without output.

wlopm is another cli that use wlr output power management instead, which
power on/off the outputs.

This protocol is old and well supported by Sway and River.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>