i3 input handling switch from bindsym to xbindkeys
This patch replaces the work in commit 489956e3 and introduces xbindkeys as a
new dependancy.

Using bindsym it was discovered with lengthy testing that the no input bug that
sway has appeared in i3.

Switching to xbindkeys has prevented this issue from reoccurring.

Unfortunately I was unable to deduce what the root cause was.

Signed-off-by: hazardchem <hazardchem@disroot.org>