Fix audio profile setup with latest Wireplumber
Wireplumber release 0.5.11 now include a hook to automatically switch
the audio profile when an active call start and stop. This conflict
directly with callaudio if it also tries to switch profiles too.

Callaudio has been patched to detect this, and to not switch profile
itself when it detect Wireplumber. Making it harmless. But because we
also support Pulseaudio, and because the Pinephone use a different
callaudio backend, we still have to adapt how we setup the call audio
mode. (also because it cost resources to call callaudiod when it is
pointless).

In the first place, this commit effectively strip all features that
cannot be covered by both approaches. We was using callaudiocli to
control the mic and the speaker mode by example. We was also displaying
the status bar icons in green when an audio profile is set. This strip
everything, including the hook. We will have to talk to the audio
daemon directly from now.

We guess we are using Wireplumber when using Pipewire. If someone know
a safer way to do that \o/