Prevent broken superd or pipewire on Super+e
We restart superd too quickly, before the previous instance effectively
stop.

This take time to cleanup the whole cleanup process. Previously there
was a cleanup() method for every supported WM. River was already
using the common scripts to close the virtual keyboard, and menu. So
we can move this to the common script. This also centralize the 5 second
cooldown before finishing, so that some service can correctly stop
themselves before tinydm restarts Sxmo.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>