sxmo_wm.sh: search for handlers from most specific to least
This makes it easier to override the behavior for specific wms.

To ensure we don't accidentally match external commands such as "paste",
prefix generic handlers with "wm_generic_".

We also need to rename some of the helper functions for focusedwindow so
that they don't take priority over the generic implementation now.

Signed-off-by: Aren Moynihan <aren@peacevolution.org>