sxmo_wm.sh: refactor execwait to avoid polling and combine common logic
By using a fifo to signal the waiting process when the command has
exited, we can avoid needing a pidfile and polling to see if it's still
alive.

While testing this, I found that `swaymsg exec` performs shell expansion
on the parameters passed to it, this attempts to bypass that to match
the behavior or dwm as I believe that is the less surprising method.
That is a breaking change.

Sway (and i3) also interpret arguments passed to them in a way that's a
bit different from shell escaping (I haven't dug into how exactly) so
the command is written to a file to bypass that.

I've added a test case for this, which will be introduced later in this
series.

Signed-off-by: Aren Moynihan <aren@peacevolution.org>