sxmo_wm.sh: rename xorg actions using xdotool to be dwm specific
These rely on window manager keybindings, which are likely to be
different on other wms. The xorg name is from when we only supported
dwm, but since we've added support for i3 (and potentially more in the
future) we should be clear about which wms they're expected to work in.

Signed-off-by: Aren Moynihan <aren@peacevolution.org>