sxmo_idle.sh: refactoring, bug removal, performance improvements
This patch adds a man page for sxmo_idle.sh and in the script it has a help function.

Also found room for improvement in the _swayidletoidles function which allowed
easier parsing in the xorgidle function, reducing the need of piping commands around.
There was a bug in the original script because trap included EXIT it would run the
finish script twice when using Ctrl+C. Using a second trap only for exit in the
cleanup function resolves this issue.

Signed-off-by: hazardchem <hazardchem@disroot.org>